move loadSrc to utils and use in process-style
