+++ lib/process-style.js
@@ -6,6 +6,7 @@ const compileStyle = require('@vue/component-compiler-utils').compileStyle
 const applyModuleNameMapper = require('./module-name-mapper-helper')
 const getCustomTransformer = require('./utils').getCustomTransformer
 const logResultErrors = require('./utils').logResultErrors
+const loadSrc = require('./utils').loadSrc
 
 function getGlobalResources(resources, lang) {
   let globalResources = ''
@@ -48,15 +49,11 @@ function getPreprocessOptions(lang, filePath, jestConfig) {
   }
 }
 
-module.exports = function processStyle(stylePart, filename, config = {}) {
+module.exports = function processStyle(stylePart, filePath, config = {}) {
   const vueJestConfig = getVueJestConfig(config)
 
   if (stylePart.src && !stylePart.content) {
-    const styleSrcPath = path.resolve(
-      filename.substring(0, filename.lastIndexOf('/')),
-      stylePart.src
-    )
-    stylePart.content = fs.readFileSync(styleSrcPath)
+    stylePart.content = loadSrc(stylePart.src, filePath)
   }
 
   if (vueJestConfig.experimentalCSSCompile === false || !stylePart.content) {
@@ -72,21 +69,21 @@ module.exports = function processStyle(stylePart, filename, config = {}) {
 
   // pre process
   if (transformer.preprocess) {
-    content = transformer.preprocess(content, filename, config, stylePart.attrs)
+    content = transformer.preprocess(content, filePath, config, stylePart.attrs)
   }
 
   // transform
   if (transformer.process) {
-    content = transformer.process(content, filename, config, stylePart.attrs)
+    content = transformer.process(content, filePath, config, stylePart.attrs)
   } else {
     const preprocessOptions = getPreprocessOptions(
       stylePart.lang,
-      filename,
+      filePath,
       config
     )
     const result = compileStyle({
       source: content,
-      filename,
+      filePath,
       preprocessLang: stylePart.lang,
       preprocessOptions,
       scoped: false
@@ -97,7 +94,7 @@ module.exports = function processStyle(stylePart, filename, config = {}) {
 
   // post process
   if (transformer.postprocess) {
-    return transformer.postprocess(content, filename, config, stylePart.attrs)
+    return transformer.postprocess(content, filePath, config, stylePart.attrs)
   }
 
   return JSON.stringify(extractClassMap(content))
