+++ lib/process.js
@@ -5,13 +5,11 @@ const generateSourceMap = require('./generate-source-map')
 const typescriptTransformer = require('./typescript-transformer')
 const coffeescriptTransformer = require('./coffee-transformer')
 const _processStyle = require('./process-style')
-const fs = require('fs')
-const path = require('path')
 const getVueJestConfig = require('./utils').getVueJestConfig
 const logResultErrors = require('./utils').logResultErrors
 const stripInlineSourceMap = require('./utils').stripInlineSourceMap
 const getCustomTransformer = require('./utils').getCustomTransformer
-const throwError = require('./utils').throwError
+const loadSrc = require('./utils').loadSrc
 const babelTransformer = require('babel-jest')
 const compilerUtils = require('@vue/component-compiler-utils')
 const convertSourceMap = require('convert-source-map')
@@ -25,18 +23,6 @@ function resolveTransformer(lang = 'js', vueJestConfig) {
     return transformer || coffeescriptTransformer
   } else {
     return transformer || babelTransformer
-  }
-}
-
-function loadSrc(src, filePath) {
-  var dir = path.dirname(filePath)
-  var srcPath = path.resolve(dir, src)
-  try {
-    return fs.readFileSync(srcPath, 'utf-8')
-  } catch (e) {
-    throwError(
-      'Failed to load src: "' + src + '" from file: "' + filePath + '"'
-    )
   }
 }
