+++ lib/utils.js
@@ -2,6 +2,7 @@ const loadPartialConfig = require('@babel/core').loadPartialConfig
 const createTransformer = require('ts-jest').createTransformer
 const chalk = require('chalk')
 const path = require('path')
+const fs = require('fs')
 
 const fetchTransformer = function fetchTransformer(key, obj) {
   for (const exp in obj) {
@@ -114,6 +115,18 @@ const logResultErrors = result => {
   }
 }
 
+const loadSrc = (src, filePath) => {
+  var dir = path.dirname(filePath)
+  var srcPath = path.resolve(dir, src)
+  try {
+    return fs.readFileSync(srcPath, 'utf-8')
+  } catch (e) {
+    throwError(
+      'Failed to load src: "' + src + '" from file: "' + filePath + '"'
+    )
+  }
+}
+
 module.exports = {
   stripInlineSourceMap,
   throwError,
@@ -126,5 +139,6 @@ module.exports = {
   info,
   warn,
   resolvePath,
-  fetchTransformer
+  fetchTransformer,
+  loadSrc
 }
